move preview component
